master sync

+++ Graph/PowerShell_SDK/AssignedLicensesSDK.ps1
@@ -1,6 +1,6 @@
 <#
 Name......: AssignedLicensesSDK.ps1
-Version...: 24.07.2
+Version...: 25.03.2
 Author....: Dario CORRADA
 
 This script will connect to the Microsoft 365 tenant and query a list of which 
@@ -9,11 +9,9 @@ license(s) are assigned to each user, then create/edit an excel report file.
 Thx to Ali TAJRAN for the useful notes about Get-MgUser on:
 https://www.alitajran.com/get-mguser/ 
 
-+++ TO DO +++
-* Integrate a feature which produce summary aomunt report of individual licenses 
-  assigned vs. that ones recovered (aka available to be assigned again), based 
-  on a specific time interval (ie monthly). See also the "MOTA.ps1" proof of 
-  concept on tempus branch.
+TODO LIST:
+* to build new pivot template 
+  (happy ending option reserved for initial Excel reference files)
 #>
 
 
@@ -517,18 +515,69 @@ try {
 }
 $ErrorActionPreference= 'Inquire'
 
+<# MOTA add-on
+Add a worksheet which produce summary amount report of individual licenses 
+assigned vs. that ones recovered (aka available to be assigned again).
+See also the "/AzureAD/MOTA.ps1" script for more details.
+#>
+Close-ExcelPackage -ExcelPackage $XlsPkg
+
+# remove older worksheet
+if ($Worksheet_list.Name -contains 'MOTA') {
+    Remove-Worksheet -Path $xlsx_file -WorksheetName 'MOTA'
+    Write-Host -NoNewline "Updating worksheet [MOTA]..."
+} else {
+    Write-Host -NoNewline "Adding worksheet [MOTA]..."
+}
+
+# launch MOTA and import related csv file
+$MOTAcsvfile = PowerShell.exe -file "$workdir\AzureAD\MOTA.ps1" -InFile $xlsx_file
+# by default $MOTAcsvfile should be "C:$($env:HOMEPATH)\Downloads\mota.csv"
+
+$inData = Import-Csv $MOTAcsvfile | ForEach-Object {
+    Write-Host -NoNewline '.'        
+    New-Object -TypeName PSObject -Property @{
+        TIMESTAMP   = [DateTime]$_.TIMESTAMP
+        ACCOUNT     = $_.ACCOUNT
+        LICENSE     = $_.LICENSE
+        STATUS      = $_.STATUS
+    } | Select TIMESTAMP, ACCOUNT, LICENSE, STATUS
+}
+$XlsPkg = Open-ExcelPackage -Path $xlsx_file
+$XlsPkg = $inData | Export-Excel -ExcelPackage $XlsPkg -WorksheetName 'MOTA' -TableName 'MOTA' -TableStyle 'Medium4' -AutoSize -PassThru
+Write-Host -ForegroundColor Green ' DONE'
+
 # resorting worksheets
 $XlsPkg.Workbook.Worksheets.MoveToStart('SkuCatalog')
 $XlsPkg.Workbook.Worksheets.MoveAfter('Licenses_Pool', 'Skucatalog')
 $XlsPkg.Workbook.Worksheets.MoveAfter('Assigned_Licenses', 'Licenses_Pool')
 if ($XlsPkg.Workbook.Worksheets.Name -contains 'Orphaned') {
     $XlsPkg.Workbook.Worksheets.MoveAfter('Orphaned', 'Assigned_Licenses')
+    $XlsPkg.Workbook.Worksheets.MoveAfter('MOTA', 'Orphaned')
+} else {
+    $XlsPkg.Workbook.Worksheets.MoveAfter('MOTA', 'Assigned_Licenses')
 }
 
-# brand new pivot example for freshly new excel files
-if ($UseRefFile -eq 'No') { 
+<# *******************************************************************************
+                                HAPPY ENDING
+******************************************************************************* #>
+$HappyEnding = @{}
+$form_panel = FormBase -w 300 -h 220 -text "HAPPY ENDING"
+if ($UseRefFile -eq 'No') {
+    # enable this feature once implemented the new release
+    $HappyEnding['PIVOT'] = CheckBox -form $form_panel -checked $false -enabled $false -x 50 -y 20 -text "Add summary pivot template"
+} else {
+    $HappyEnding['PIVOT'] = CheckBox -form $form_panel -checked $false -enabled $false -x 50 -y 20 -text "Add summary pivot template"
+}
+$HappyEnding['PREVIEW'] = CheckBox -form $form_panel -checked $true -x 50 -y 50 -text "Open Excel formatted file"
+$HappyEnding['CLEANSWEEP'] = CheckBox -form $form_panel -checked $false -x 50 -y 80 -text "Remove temporary backup"
+OKButton -form $form_panel -x 80 -y 130 -text "Ok"  | Out-Null
+$result = $form_panel.ShowDialog()
+
+if ($HappyEnding['PIVOT'].Checked -eq $true) {
+<# review needed
     Add-Worksheet -ExcelPackage $XlsPkg -WorksheetName 'SUMMARY' > $null
-    
+
     Add-PivotTable -ExcelPackage $XlsPkg `
     -PivotTableName 'POOL' -Address $XlsPkg.SUMMARY.cells["B3"] `
     -SourceWorksheet 'Licenses_Pool' `
@@ -541,21 +590,17 @@ if ($UseRefFile -eq 'No') {
     -SourceWorksheet 'Assigned_Licenses' `
     -PivotRows ('LICENSE', 'DESC') -PivotColumns 'TIMESTAMP' -PivotData 'LICENSE' `
     -PivotTableStyle 'Dark3' -PivotTotals 'Rows'
+#> 
 }
 
-# show the final result and/or keep temporary backup
-$ErrorActionPreference= 'Stop'
-try {
+if ($HappyEnding['PREVIEW'].Checked -eq $true) {
     Close-ExcelPackage -ExcelPackage $XlsPkg -Show
-} catch {
+} else {
     Close-ExcelPackage -ExcelPackage $XlsPkg
 }
-$ErrorActionPreference= 'Inquire'
-if ($UseRefFile -eq 'Yes') {
-    $answ = [System.Windows.MessageBox]::Show("Remove temporary backup?",'DELETE','YesNo','Warning')
-    if ($answ -eq "Yes") {    
-        Remove-Item -Path $bkp_file -Force
-    }
+
+if (($HappyEnding['CLEANSWEEP'].Checked -eq $true) -and ($UseRefFile -eq 'Yes'))  {
+    Remove-Item -Path $bkp_file -Force
 }
 
 # REMOVE RENMANTS
